Worked on all components and css. App.css and Resume.css need completion
