+++ portfolio/src/App.css
@@ -22,7 +22,7 @@ body {
   background-image: linear-gradient(to right, grey, white);
 }
 
-main {
+/* main {
   margin-top: -50px;
 }
 
@@ -41,21 +41,21 @@ a {
 
 a:visited {
   /* Previously clicked */
-  color: #502846;
-}
+  /* color: #502846; */
+/* } */
 
-a:hover,
-a:focus {
+/* a:hover, */
+/* a:focus { */
   /* No underlining when hovering over a link */
-  text-decoration: none;
-}
+  /* text-decoration: none; */
+/* } */
 
-a:active {
+/* a:active { */
   /* Selected link */
-  color: gray;
-}
+  /* color: gray; */
+/* } */
 
-.profile-description {
+/* .profile-description {
   background-color: black;
   color: darkgoldenrod;
   padding: 0 2px 2px 2px;
@@ -86,8 +86,6 @@ a:active {
   height: 200px;
 }
 
-
-
 .easter-egg {
   opacity: 0;
   position: absolute;
@@ -96,11 +94,11 @@ a:active {
 .easter-egg:hover {
   opacity: 1;
   cursor: pointer;
-}
+} */
 
 /* MEDIA QUERIES */
 
-@media (min-width: 875px) and (max-width: 1000px) {
+/* @media (min-width: 875px) and (max-width: 1000px) {
   h1 {
     font-size: 30px;
   }
@@ -192,4 +190,4 @@ a:active {
     height: auto;
     margin-right: 50px;
   }
-}
+} */
