+++ portfolio/src/components/Footer/Footer.css
@@ -3,6 +3,7 @@
   bottom: 20px;
   width: 100%;
   text-align: center;
+  margin-bottom: 25px;
 }
 
 .social-links {
