+++ portfolio/src/components/Home/Home.css
@@ -1,0 +1,117 @@
+main {
+  margin-top: -50px;
+}
+
+h1 {
+  font-size: 35px;
+}
+
+h2 {
+  font-size: 24px;
+  margin-left: 15px;
+}
+
+.profile-description {
+  background-color: black;
+  color: darkgoldenrod;
+  padding: 0 2px 2px 2px;
+}
+
+.profile__portrait {
+  float: left;
+  width: 250px;
+  height: auto;
+  margin-right: 100px;
+  border-radius: 50%;
+  border: solid gray;
+}
+
+.profile {
+  max-width: 700px;
+  margin: 100px auto;
+}
+
+.projects {
+  padding: 1em;
+  text-align: center;
+}
+
+.project-item {
+  border: solid gray;
+  width: 200px;
+  height: 200px;
+}
+
+.easter-egg {
+  opacity: 0;
+  position: absolute;
+}
+
+.easter-egg:hover {
+  opacity: 1;
+  cursor: pointer;
+}
+
+/* MEDIA QUERIES */
+
+@media (min-width: 875px) and (max-width: 1000px) {
+  h1 {
+    font-size: 30px;
+  }
+
+  .profile__portrait {
+    float: left;
+    width: 220px;
+    height: auto;
+    margin-right: 50px;
+  }
+}
+
+@media (min-width: 650px) and (max-width: 874px) {
+  body {
+    font-size: 18px;
+  }
+
+  h1 {
+    font-size: 28px;
+  }
+
+  .profile__portrait {
+    float: left;
+    width: 200px;
+    height: auto;
+    margin-right: 50px;
+  }
+}
+
+@media (max-width: 649px) {
+  body {
+    font-size: 16px;
+  }
+
+  h1 {
+    font-size: 25px;
+  }
+
+  h2 {
+    font-size: 20px;
+  }
+
+  .profile {
+    display: block;
+    text-align: center;
+    margin-bottom: auto;
+  }
+
+  .profile__portrait {
+    float: none;
+    width: 175px;
+    height: auto;
+    margin-right: auto;
+    margin-top: -70px;
+  }
+
+  .projects {
+    margin-bottom: -30px;
+  }
+}
