--- portfolio/src/components/Resume.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Resume = () => {
-  return (
-    <div>
-      <h1>Resume Component</h1>
-    </div>
-  );
-};
-
-export default Resume;
