+++ portfolio/src/components/Resume/Resume.css
@@ -1,0 +1,8 @@
+.resume-img {
+  display: block;
+  margin-top: 40px;
+  margin-bottom: 40px;
+  margin-left: auto;
+  margin-right: auto;
+  width: 50%;
+}
